# frontend/lib/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DR-OS Control Plane",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/uploads/sign": {
      "post": {
        "tags": [
          "gateway"
        ],
        "summary": "Sign Upload",
        "operationId": "sign_upload_v1_uploads_sign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignedUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/uploads/complete": {
      "post": {
        "tags": [
          "gateway"
        ],
        "summary": "Complete Upload",
        "operationId": "complete_upload_v1_uploads_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadCompleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/events": {
      "get": {
        "tags": [
          "gateway"
        ],
        "summary": "Stream Project Events",
        "operationId": "stream_project_events_v1_projects__project_id__events_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "once",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Once"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/artifacts/{artifact_id}/download-url": {
      "get": {
        "tags": [
          "gateway"
        ],
        "summary": "Get Artifact Download Url",
        "operationId": "get_artifact_download_url_v1_projects__project_id__artifacts__artifact_id__download_url_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedArtifactUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project",
        "operationId": "update_project_v1_projects__project_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{project_id}/manuscripts": {
      "post": {
        "tags": [
          "manuscripts"
        ],
        "summary": "Create Manuscript",
        "operationId": "create_manuscript_v1_projects__project_id__manuscripts_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManuscriptRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateManuscriptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "manuscripts"
        ],
        "summary": "List Manuscripts",
        "operationId": "list_manuscripts_v1_projects__project_id__manuscripts_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManuscriptListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/manuscripts/{manuscript_id}": {
      "get": {
        "tags": [
          "manuscripts"
        ],
        "summary": "Get Manuscript",
        "operationId": "get_manuscript_v1_projects__project_id__manuscripts__manuscript_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "manuscript_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Manuscript Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManuscriptDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/manuscripts/{manuscript_id}/blocks": {
      "post": {
        "tags": [
          "manuscripts"
        ],
        "summary": "Create Block",
        "operationId": "create_block_v1_projects__project_id__manuscripts__manuscript_id__blocks_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "manuscript_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Manuscript Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManuscriptBlockRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateManuscriptBlockResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "manuscripts"
        ],
        "summary": "List Blocks",
        "operationId": "list_blocks_v1_projects__project_id__manuscripts__manuscript_id__blocks_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "manuscript_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Manuscript Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManuscriptBlockListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/manuscripts/{manuscript_id}/versions": {
      "post": {
        "tags": [
          "manuscripts"
        ],
        "summary": "Create Version",
        "operationId": "create_version_v1_projects__project_id__manuscripts__manuscript_id__versions_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "manuscript_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Manuscript Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManuscriptVersionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateManuscriptVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/manuscripts/{manuscript_id}/render": {
      "post": {
        "tags": [
          "manuscripts"
        ],
        "summary": "Render Manuscript",
        "operationId": "render_manuscript_v1_projects__project_id__manuscripts__manuscript_id__render_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "manuscript_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Manuscript Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderManuscriptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlockType": {
        "type": "string",
        "enum": [
          "text",
          "figure",
          "table",
          "citation_list"
        ],
        "title": "BlockType"
      },
...
      "ComplianceLevel": {
        "type": "string",
        "enum": [
          "public",
          "internal",
          "clinical"
        ],
        "title": "ComplianceLevel"
      },
...
      "CreateManuscriptBlockRequest": {
        "properties": {
          "section_key": {
            "$ref": "#/components/schemas/SectionKey"
          },
          "block_order": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Block Order",
            "default": 0
          },
          "block_type": {
            "$ref": "#/components/schemas/BlockType",
            "default": "text"
          },
          "content_md": {
            "type": "string",
            "minLength": 1,
            "title": "Content Md"
          },
          "assertion_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Assertion Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "section_key",
          "content_md"
        ],
        "title": "CreateManuscriptBlockRequest"
      },
      "CreateManuscriptBlockResponse": {
        "properties": {
          "block": {
            "$ref": "#/components/schemas/ManuscriptBlockRead"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "block"
        ],
        "title": "CreateManuscriptBlockResponse"
      },
      "CreateManuscriptRequest": {
        "properties": {
          "manuscript_type": {
            "$ref": "#/components/schemas/ManuscriptType",
            "default": "manuscript"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "target_journal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Journal"
          },
          "style_profile_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Style Profile Json"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CreateManuscriptRequest"
      },
      "CreateManuscriptResponse": {
        "properties": {
          "manuscript": {
            "$ref": "#/components/schemas/ManuscriptRead"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "manuscript"
        ],
        "title": "CreateManuscriptResponse"
      },
      "CreateManuscriptVersionRequest": {
        "properties": {
          "base_version_no": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Version No"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CreateManuscriptVersionRequest"
      },
      "CreateManuscriptVersionResponse": {
        "properties": {
          "manuscript": {
            "$ref": "#/components/schemas/ManuscriptRead"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "manuscript"
        ],
        "title": "CreateManuscriptVersionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManuscriptBlockListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ManuscriptBlockRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ManuscriptBlockListResponse"
      },
...
      "ManuscriptDetailResponse": {
        "properties": {
          "manuscript": {
            "$ref": "#/components/schemas/ManuscriptRead"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "manuscript"
        ],
        "title": "ManuscriptDetailResponse"
      },
      "ManuscriptListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ManuscriptRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ManuscriptListResponse"
      },
...
      "ManuscriptType": {
        "type": "string",
        "enum": [
          "manuscript",
          "abstract",
          "grant_response"
        ],
        "title": "ManuscriptType"
      },
...
      "ProjectDetailResponse": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectRead"
          },
          "active_workflows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowInstanceRead"
            },
            "type": "array",
            "title": "Active Workflows"
          },
          "latest_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetSnapshotRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_manuscript": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManuscriptRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Review Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project"
        ],
        "title": "ProjectDetailResponse"
      },
...
      "ProjectState": {
        "type": "string",
        "enum": [
          "draft",
          "running",
          "review_required",
          "approved",
          "archived"
        ],
        "title": "ProjectState"
      },
...
      "RenderManuscriptResponse": {
        "properties": {
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/ManuscriptBlockRead"
            },
            "type": "array",
            "title": "Blocks"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "blocks"
        ],
        "title": "RenderManuscriptResponse"
      },
...
      "SectionKey": {
        "type": "string",
        "enum": [
          "title",
          "abstract",
          "introduction",
          "methods",
          "results",
          "discussion",
          "conclusion",
          "figure_legend",
          "table_note",
          "appendix"
        ],
        "title": "SectionKey"
      },
...
      "SignedArtifactUrlResponse": {
        "properties": {
          "download_url": {
            "type": "string",
            "title": "Download Url"
          },
          "expires_in_seconds": {
            "type": "integer",
            "title": "Expires In Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "download_url",
          "expires_in_seconds"
        ],
        "title": "SignedArtifactUrlResponse"
      },
      "SignedUploadRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "minLength": 1,
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "minLength": 1,
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size Bytes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filename",
          "content_type",
          "size_bytes"
        ],
        "title": "SignedUploadRequest"
      },
      "SignedUploadResponse": {
        "properties": {
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "expires_in_seconds": {
            "type": "integer",
            "title": "Expires In Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "upload_url",
          "object_key",
          "expires_in_seconds"
        ],
        "title": "SignedUploadResponse"
      },
...
      "UpdateProjectRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectState"
              },
              {
                "type": "null"
              }
            ]
          },
          "compliance_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComplianceLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_manuscript_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Manuscript Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UpdateProjectRequest"
      },
      "UploadCompleteRequest": {
        "properties": {
          "object_key": {
            "type": "string",
            "minLength": 1,
            "title": "Object Key"
          },
          "sha256": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "title": "Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "object_key",
          "sha256"
        ],
        "title": "UploadCompleteRequest"
      },
      "UploadCompleteResponse": {
        "properties": {
          "file_ref": {
            "type": "string",
            "title": "File Ref"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "file_ref"
        ],
        "title": "UploadCompleteResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}